feat(condo): DOMA-13104 added planType filter to GetAvailablePlansService
